# appinio-backend/output/openapi.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"getHelloFromAppController","parameters":[],"responses":{"200":{"description":""}}}},"/posts":{"get":{"operationId":"findAllFromPostsController","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostDto"}}}}}}},"post":{"operationId":"createFromPostsController","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePostRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostDto"}}}},"201":{"description":""}}}},"/posts/summarize":{"post":{"operationId":"summarizeTextFromPostsController","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SummarizePostRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SummarizePostResponseDto"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SummarizePostResponseDto"}}}}}}},"/auth/login":{"post":{"operationId":"signInFromAuthController","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInResponseDto"}}}}}}},"/auth/register":{"post":{"operationId":"registerFromAuthController","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInResponseDto"}}}}}}}},"info":{"title":"Appinio API","description":"Appinio API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"PostDto":{"type":"object","properties":{"id":{"type":"number"},"uuid":{"type":"string"},"content":{"type":"string"},"summary":{"type":"string"},"insights":{"type":"string"},"userId":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","uuid","content","summary","insights","userId","createdAt","updatedAt"]},"CreatePostRequestDto":{"type":"object","properties":{"content":{"type":"string"},"summary":{"type":"string"},"insights":{"type":"string"}},"required":["content","summary","insights"]},"SummarizePostRequestDto":{"type":"object","properties":{"content":{"type":"string"}},"required":["content"]},"SummarizePostResponseDto":{"type":"object","properties":{"summary":{"type":"string"},"insights":{"type":"string"}},"required":["summary","insights"]},"SignInRequestDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"SignInResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"}},"required":["accessToken","username","email"]},"RegisterRequestDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"}},"required":["username","password","email"]}}}}